Attempting (noting that I know nothing of how to use Python) to integrate the functionality I producted in Octave whereby a CombateFlite NS430 data file can be imported directly into wpted.  Right now the Octave script is here: https://github.com/gavinedwards/cf2wpe
I have got as far as figuring out how to create an event(?) in the Profile -> Import menu called Import from CombatFlite NS430 Data.  It is duplicated from the Load from Encoded File function right now.  Next step I think is to define the function to replicate the functionality of the Octave script.